LayerEditor refactor
