+++ src/components.d.ts
@@ -10,17 +10,16 @@ declare module 'vue' {
     AppFooter: typeof import('./components/AppFooter.vue')['default']
     BoundaryLine: typeof import('./components/BoundaryLine.vue')['default']
     ConfirmDelete: typeof import('./components/ConfirmDelete.vue')['default']
-    copy: typeof import('./components/LayerEditor copy.vue')['default']
     EarthAnimation: typeof import('./components/misc/EarthAnimation.vue')['default']
     FindClosest: typeof import('./components/FindClosest.vue')['default']
     GameCreator: typeof import('./components/game/GameCreator.vue')['default']
     GameSelector: typeof import('./components/game/GameSelector.vue')['default']
     HelloWorld: typeof import('./components/HelloWorld.vue')['default']
     History: typeof import('./components/History.vue')['default']
-    LayerEditor: typeof import('./components/LayerEditor.vue')['default']
+    LayerEditor: typeof import('./components/dialog/LayerEditor.vue')['default']
     Map: typeof import('./components/Map.vue')['default']
     MapActionButton: typeof import('./components/MapActionButton.vue')['default']
-    MarkerEditor: typeof import('./components/MarkerEditor.vue')['default']
+    MarkerEditor: typeof import('./components/dialog/MarkerEditor.vue')['default']
     Radar: typeof import('./components/Radar.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
