+++ src/components/dialog/LayerEditor.vue
@@ -1,0 +1,71 @@
+<template>
+  <v-dialog max-width="350" v-model="model as any" transition="dialog-bottom-transition">
+    <v-card title="Choose Map Layers">
+      <v-card-text>
+        <v-alert style="margin-bottom: 1em;" icon="$info" density="compact">
+          This will only apply to your device.
+        </v-alert>
+        <v-radio-group label="Background style" v-model="mapStyle" v-on:click="updateMap" >
+          <v-radio label="Dark" value="dark"></v-radio> <!-- Jawg.Matrix or Jawg.Dark -->
+          <v-radio label="Light" value="light"></v-radio> <!-- Jawg.Sunny  -->
+          <v-radio label="Terrain" value="terrain"></v-radio> <!-- Esri_WorldImagery  -->
+        </v-radio-group>
+        <v-checkbox label="Railroad Overlay" v-model="enableRailroadOverlay" v-on:update:model-value="updateMap"></v-checkbox>
+      </v-card-text>
+      <v-card-actions>
+        <v-container>
+          <v-row align="center" justify="center">
+            <v-col cols="12" md="6">
+              <v-btn prepend-icon="mdi-close" variant="tonal" @click="model = false" block>Close</v-btn>
+            </v-col>
+          </v-row>
+        </v-container>
+      </v-card-actions>
+    </v-card>
+  </v-dialog>
+</template>
+
+<script lang="ts" setup>
+import { useStore } from '@/stores/app';
+
+const store = useStore();
+
+const model = defineModel()
+
+const mapStyle = shallowRef("dark");
+const enableRailroadOverlay = shallowRef(false);
+
+const updateMap = async() => {
+  await new Promise(r => setTimeout(r, 200));
+  const newLayers = [];
+  switch (mapStyle.value) {
+    case "dark": newLayers.push("Jawg.Matrix"); break;
+    case "light": newLayers.push("Jawg.Sunny"); break;
+    case "terrain": newLayers.push("Esri_WorldImagery"); break;
+  }
+
+  if (enableRailroadOverlay.value) {
+    newLayers.push("OpenRailwayMap");
+  }
+
+  store.$state.mapLayers = newLayers;
+};
+
+// Refresh on every view - store could be edited elsewhere.
+watch(model, () => {
+  if (model.value) {
+    if (store.$state.mapLayers.includes("Jawg.Matrix")) {
+      mapStyle.value = "dark";
+    } else if (store.$state.mapLayers.includes("Jawg.Sunny")) {
+      mapStyle.value = "light";
+    } else if (store.$state.mapLayers.includes("Esri_WorldImagery")) {
+      mapStyle.value = "terrain";
+    }
+
+    if (store.$state.mapLayers.includes("OpenRailwayMap")) {
+      enableRailroadOverlay.value = true;
+    }
+  }
+})
+
+</script>
